+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [12.17.0] - 2024-07-21
+### Improved
+- **feat**: Barrel export for array now also includes new split function
+```
+import {split} from '@valkyriestudios/utils/array';
+```
+
 ## [12.16.0] - 2024-07-21
 ### Added
 - **feat**: array/split utility function which can be used to split an array into batches of a particular size